pyproject.toml
[project]
name = "kiero"
version = "0.1.0"
description = "Manga watermark detector and remover"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "opencv-python",
    "Pillow",
    "torch",
    "torchvision",
    "ultralytics",
    "transformers",
    "simple-lama-inpainting",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
kiero = "kiero.cli:main"
